Output: Generate CvQualifiedType elements
Implement the OutputCvQualifiedType method to generate an element whose
ID attribute id="" is that of an unqualified type plus 'c', 'v', or 'r'
letters appended to indicate qualification.  Reference the IDREF of the
unqualified type in a type="" attribute and describe the qualifiers with
const="1", volatile="1", and restrict="1" attributes as appropriate.