Output: Generate FunctionType elements
Implement the OutputFunctionProtoType method to generate the
FunctionType element.  Report the return type.  Inside the FunctionType
element generate Argument elements describing the function type
parameter types.  For each argument report the type and queue it.

Create a OutputFunctionTypeHelper method to hold the main implementation
and call it from OutputFunctionProtoType with the "FunctionType" name as
an argument.  This will allow it to be re-used for other function types
later.