Output: Extend type nodes and queue to pair with containing class
Clang represents pointer-to-member-function as a pointer-to-member type
whose pointee is a normal function-type.  The gccxml output format
represents them as a normal pointer whose pointee is a "method-type".
Since clang::QualType cannot represent a "method-type", introduce our
own DumpType struct to contain a clang::QualType and a clang::Type*
where the latter can be used to point to the containing class.