test: Show duplication of reference-to-class-template-specialization types
Clang generates a distinct TemplateSpecializationType for each
appearance of A<T> for some template A<>.  They are all sugar for an
underlying RecordType.  While they all share the same canonical type,
LValueReferenceType instances that refer to them are also distinct and
do not share the same canonical type because each has a different
pointee type.

Add tests demonstrating cases where this causes duplicate ReferenceType
elements to be generated.  Mark them as "broken" and add a note in the
expected output that matches nothing but documents the problem.