Output: Order queue deterministically by node index
Use a set<> instead of a queue<> to represent the output node queue.
Order nodes by the dump node index.  The index is the same as the order
of encounter so for complete nodes this should normally be the same
order as before.  However, QueueIncompleteDumpNodes previously ordered
based on memory position which is not deterministic.  Use the node index
to get a consistent order across runs.  Fix the expected output for test
cases that changed the order of incomplete nodes.