RunClang: With --castxml-cc-<id>, disable Clang stdlib includes
Tell Clang not to add its own standard library include directories when
we are simulating another compiler.  Pass -nobuiltininc and -nostdlibinc
to suppress all directories that Clang might add.  Do not pass -nostdinc
because that may prevent Clang code from ever checking for -nostdlibinc
and then it would warn that the argument is not used.