RunClang: Delay class implicit members until end of parsing
Tell Clang to enable incremental processing so that it does not tear
down the parser when EOF is reached.  In HandleTagDeclDefinition, queue
completed classes for later processing instead of adding class implicit
members immediately.  Then add implicit members at the end and finally
call clang::Sema::ActOnEndOfTranslationUnit to tear down the parser when
we are done.  This approach allows Clang to finish parsing the
translation unit without interference from our custom implicit member
behavior.

Fix the expected output of the Class-bases test in c++11 mode to match
the new order of the implicit members.