Output: Fix references to cv-qualified types encountered indirectly
Teach AddDumpNode to optionally report back the actual QualType that it
selects after possibly unwrapping some layers of indirection in the
Clang AST.  Use this in GetTypeIdRef so that we check this actual
QualType's cv qualifiers instead of the original QualType.  Otherwise
we may generate a dangling reference to the id number that is replaced
by a CvQualifiedType id.

Fix the expected output of existing test cases that exhibit this problem
and add a new test case designed to cover the behavior explicitly.