allow --castxml-cc-X to be separated from its argument by an equal sign
  --castxml-cc-gnu=/usr/bin/clang++

With the current syntax, castxml can't be used by ccache, because castxml thinks
that /usr/bin/clang++ is an input of the compiler. With this change, the compiler
path is in the same argument than the option and thus no considered by ccache
as a part of the option, not as an input.