Detect: Do not use Clang builtin include dir with --castxml-cc-msvc
In commit 8d8aacd6 (Detect: Use Clang builtin include dir even with
--castxml-cc-<id>, 2015-02-04) we added the Clang builtin include dir
to the beginning of the include path detected by --castxml-cc-msvc.
However, since we use only the detected predefined macros and not
those builtin to Clang, the Clang builtin headers cannot be included
safely.  Revert that part of the commit for now.