Output: Handle function types with no prototype
A FunctionType may be either FunctionProtoType or FunctionNoProtoType.
Check that runtime downcasts to FunctionProtoType succeed before using
them.  For FunctionNoProtoType we will simply output the type "int()",
which is compatible with gccxml format (gccxml never lacked prototypes
because it always operates in C++ mode).  In CastXML we can encounter
functions with no prototype in C++ modes when traversing Clang buildins.

GitHub-Issue: 1