Output: Place inline namespace members in containing namespace
An inline namespace affects the linkage of its members but for APIs
the members are effectively in the containing namespace.  Generate
output as if they were really in the containing namespace and do
not generate a Namespace element for an inline namespace.

If --castxml-start names an inline namespace simply ignore it to avoid
dumping a Namespace element for it.  One can argue this is valid since
the inline namespace should not normally be named in API usage.

Set our printing policy to avoid showing the inline namespace component
automatically.  Note that expressions in the source code that hard-code
the inline namespace name will still be preserved and show the inline
namespace when printed (e.g. Variable init="" attribute expressions).

Suggested-by: Michka Popoff <michkapopoff@gmail.com>