Output: Generate a FundamentalType for our __float128 typedef
GNU compilers have a builtin __float128 type but Clang does not.  Since
commit fe1bb977ee (Detect: Provide __float128 when simulating GNU
compilers, 2015-02-24) we define __float128 as a typedef in the preamble
referencing a struct with the proper size and alignment.  Avoid exposing
this implementation detail to clients by generating a FundamentalType
element as a special case when our typedef is encountered.  This matches
gccxml behavior when __float128 is encountered.

GitHub-Issue: 18