Output: Further hide our __float128 struct from generated xml
Give the our __float128 struct an internal '__castxml' name so that it
can be identified during traversal of member lists.  Do not report the
struct as a member of the translation unit since the typedef will be
generated as a FundamentalType directly.  Also generate empty mangled=""
attributes for mangled names involving our __float128 struct since the
mangled names will not use the "g" that GCC would for the builtin type.

Add test cases that do not use --castxml-start.  Verify that no test
case produces '__castxml' in its output to ensure we always hide such
internal details from the generated output.

GitHub-Issue: 18