castxml: Support -std=c++14 with --castxml-gccxml
Drop the rejection of -std=c++14 and --castxml-gccxml together.  Extend
the test suite to run all gccxml output format tests in -std=c++14 mode
to verify the behavior of the mode is as close as possible to other
modes.  For the few tests that have c++11-specific expected output, use
it for c++14 expected output too.