Output: Fix references to types whose declaration is invalid
We skip generating declarations that Clang has deemed invalid.  Most
callers of AddDeclDumpNode can check the return value and handle the
lack of a valid decl.  However, callers of AddDeclDumpNodeForType are
generating references to types and expect to get a valid element to
reference.  We can get an invalid RecordDecl when it is a template
instantation that fails (but does not otherwise produce a compilation
error in Clang).  References to types corresponding to the invalid
RecordDecl must remain valid, so generate a Class/Struct/Union element
anyway and simply leave out class members and bases.