RunClang: Avoid infinite class template implicit member recursion
We force declaration of class implicit members in order to generate a
full description of valid members.  In the case that a class is a
template instantiation the definition of the implicit member bodies may
recursively create new template instantiations.  We queue them and
continue.  However, it is possible that this process will produce an
unbounded depth of template instantiations that is not diagnosed by
Clang's template instantiation depth limit because that does not force
implicit member definition.

Avoid infinite class template recursion by enforcing an artificial limit
on the depth through which we force definition of class implicit
members.

GitHub-Issue: CastXML/CastXML#50