Output: Print fully qualified declarations named in default arguments
In code like

    namespace ns {
      static int const C = 0;
      void f(int = C);
    }

we previously generated `default="C"` for the default argument of `f`
because Clang prints the expression as it was written in the source.
GCC-XML would generate the fully qualified name `default="ns::C"`
which is more useful to tools reading the output.  Update our default
argument printing to do this by hooking in to Clang's pretty printer
and providing a custom implementation for `DeclRefExpr`.

GitHub-Issue: CastXML/CastXML#51