Output: Print canonical form of types named in default arguments
In code like

    namespace ns {
      typedef unsigned int UI;
      void f(UI = (UI)0);
    }

we previously generated `default="(UI)0"` for the default argument of
`f` because Clang prints the expression as it was written in the source.
GCC-XML would generate the canonical typed value `default="0u"`.
Ideally we should generate `default="(ns::UI)0"` but Clang does not
offer hooks into its type printing API.  Instead we can produce a value
which is useful to tools reading the output by printing the cast with a
canonical type like `default="(unsigned int)0"`.  In the case that the
type names a struct or class the canonical type will be printed as the
fully qualified name.

GitHub-Issue: CastXML/CastXML#51