Output: Print fully qualified names in variable initializers
In code like

    namespace ns {
      static int const C = 123;
      int v1 = C;
      typedef int Int;
      Int v2 = (Int)123;
    }

we previously generated `init="C"` and `init="(Int)123"` for the
variable initializers.  Produce a value more useful to tools reading the
output by using fully qualified names and/or canonical types.  This is
the same approach we already use for function default argument
expressions.

GitHub-Issue: CastXML/CastXML#51