Output: Fix access specifier of class template instantiations
In code like

    class foo {
      template <typename> class bar {};
    };

the instantiations of `foo::bar<>` should have the same access as the
original template.  Clang generates access `AS_none` for member template
instantiations because they have no meaningful access of their own.
Previously we have misinterpreted this as `AS_public` because it is not
`AS_private` or `AS_protected`.  Instead we must detect when a class is
a template instantiation and get the access specifier from its original
template.

GitHub-Issue: CastXML/CastXML#56