RunClang: Provide `__make_integer_seq` builtin when Clang does not
Visual Studio 2015 Update 2 (cl 19.00.23918) adds a `__make_integer_seq`
builtin.  It also started using the builtin in the `<type_traits>`
header.  Clang version 3.8 and above provide this builtin but older
versions do not, so CastXML fails when built against Clang 3.6 or 3.7.
Work around this problem by adding our own implementation of the builtin
when Clang does not provide it.