castxml: Fix `__float128` simulation with `-std=gnu++{11,14}` again
The approach used in commit 1414278259 (castxml: Fix `__float128`
simulation with `-std=gnu++{11,14}`, 2016-01-25) allowed Clang to use
its own builtin `__float128` type when it is available.  However, Clang
does not really implement this builtin and issues diagnostics when it is
used in some contexts (e.g. as a data member).

Instead we need to provide our own approximation of the builtin in all
cases.  Avoid a conflict with the Clang builtin by using the
preprocessor to rename the type to `__castxml__float128`.  Then replace
this name with `__float128` during output.

GitHub-Issue: CastXML/CastXML#59