WIP: Add support for elaborated type specifiers
Elaborated type specifiers may be used to refer to a previously-declared
class name (class, struct, or union) or to a previously-declared enum name.

These type specifiers are important for C code, but are also sometimes
relavant in the context of C++ code.

This patch adds the elaborated="0|1" xml tag to variable, typedef and argument
declarations.

The output version number was updated to 1.140.

Most importantly, a test was added to demonstrate the purpose
of this change (test/input/Elaborated.cxx).