castxml: Add flag --castxml-output=<v> to activate a new output format
This will allow us to move the format forward away from gccxml
compatibility in order to add information about new C++ structures.
The new format will work with all C and C++ language modes but
still not for Objective C/C++.

Retain the `--castxml-gccxml` flag to produce XML files as similar as
possible as the ones produced by gccxml, but do not allow both formats
to be specified at the same time.  For now use the same output for both
formats except for the name of the root element.  Later the new format
can be adapted to add new features.