Output: Add support for elaborated type specifiers
Elaborated type specifiers refer to a class, struct, union, or enum name
preceded by the corresponding keyword (e.g. `struct foo`).  These type
specifiers are important for C code, but are also sometimes relevant in
the context of C++ code.

Add an `ElaboratedType` XML element to the castxml output format to
represent elaborated type specifiers.  Increment the format *major*
version number to indicate the addition of an element that will require
clients to be updated.