Check LLVM installation for Clang headers early
Sometimes users try to build CastXML on a distro with a LLVM development
package installed but not the corresponding Clang development package.
This results in a compilation failure due to missing headers.  Provide
a better/earlier error message by checking for the Clang headers as
soon as LLVM is found.

Closes: #79