RunClang: Tolerate use of GNU builtin __has_unique_object_representations
The builtin is used by the GNU type_traits header in modes aware of
C++17, so we need to provide it in CastXML.  LLVM/Clang SVN r316518
(implement __has_unique_object_representations, 2017-10-24), in
development of Clang 6, added this GNU builtin.  For older versions of
LLVM/Clang, provide a minimum implementation that at least allows
`type_traits` to parse.

Fixes: #98