test: Fix cc-gnu predefined macros on Windows
For Windows targets, GNU compilers predefine `_WIN32` and `__MINGW32__`.
CastXML uses those to set the target triple correctly when using
`--castxml-cc-gnu`.  Teach our fake GNU compiler in the test suite to
define these too.  Otherwise CastXML does not add `gnu` to the target
triple.  This causes the Windows-hosted Clang to choose `msvc` in the
default triple and enable MS compatibility, which is incorrect.