Output: Generate Comment elements for declaration doc comments
Add a `Comment` XML element to the castxml output format to represent
comments.  Include only documentation comments attached to declarations.
For each comment, reference the source file location containing it such
that clients can extract the actual comment content themselves.

Increment the format *major* version number to indicate the addition of
an element that will require clients to be updated.

Fixes: #152