Detect: Fix MSVC preprocessor settings detection on Windows 11
MSVC has no way to compile without producing the object file.
Previously we used `-FoNUL` to redirect the object file to the
null device, which `cl` opened internally as `NUL.obj`.
Windows 11 seems to treat `NUL.obj` as a real file, and only
plain `NUL` as the null device.  Opening `NUL.obj` in the
current working directory races with other CastXML processes.

Avoid this problem by putting the object file in a temporary
directory that is unique to the current process.

Fixes: #211