test: Fix test output conversion Field offset values
When using `TEST_UPDATE` to convert the actual test output to the
expected test output, convert non-zero Field `offset=""` values to a
regex matching any offset.  This makes the output robust to the ABI's
type sizes and padding.  We already do this for Base offsets in cases of
multiple inheritance.

Inspired-by: Mattias Ellert <mattias.ellert@physics.uu.se>