Output: Generate RValueReferenceType elements in castxml output format
Implement the `OutputRValueReferenceType` method to generate the
`RValueRefernceType` element in the `--castxml-output=1` format.
Report the `type=""` of the pointee and queue it.

Normally we should increment the format *major* version number to
indicate the addition of an element that will require clients to be
updated.  However, the format major version has already been incremented
by commit f1220bd4 (Output: Add ElaboratedType elements with
"qualifier=" attribute, 2022-12-10) since the last release.

Issue: #242